CEK 4

+++ resources/views/budgeting/lampiran/perwal-1_ed.blade.php
@@ -130,7 +130,7 @@
         @if($rs['koderekening']==6)
         <td class="border-rincian text_blok kanan total">{{ number_format((float)$totalpenerimaan-$totalpengeluaran,2,'.',',') }}</td>
         @elseif($rs['koderekening']==5)
-        <td class="border-rincian text_blok kanan total">{{ number_format((float)round($rs['totalrekening']),2,'.',',') }}</td>
+        <td class="border-rincian text_blok kanan total">{{ number_format(round($rs['totalrekening']),2,'.',',') }}</td>
         @else
         <td class="border-rincian text_blok kanan total">{{ number_format((float)$rs['totalrekening'],2,'.',',') }}</td>
         @endif
